pyproject.toml
[tool.poetry]
name = "multiclass-flare-prediction"
version = "0.1.0"
description = ""
authors = ["Jinsu Hong <jinsuhong.knight@gmail.com>"]
readme = "README.md"
packages = [{include = "multiclass_flare_prediction"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
numpy = "^2.2.3"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
